--- src/Layout/AppSider/index.module.css
@@ -1,6 +1,0 @@
-.app_sider .logo {
-  height: 32px;
-  background: rgba(255, 255, 255, 0.3);
-  margin: 16px;
-  cursor: pointer;
-}
